+++ src/components/Image/Image.css
@@ -2,10 +2,11 @@
   position: absolute;
   display: flex;
   justify-content: center;
-  align-items: center;
   border-radius: 0;
+  border: none;
   padding: 0;
   margin: 0;
+  overflow: hidden;
 }
 
 .imgTextInput {
@@ -23,8 +24,9 @@
 }
 
 .imgText {
-  display: block;
-  padding: 1rem;
+  height: fit-content;
+  width: max-content;
+  padding: 0;
   position: absolute;
   bottom: 2rem;
   background: none;
