+++ src/components/Tooltip/Tooltip.css
@@ -1,0 +1,63 @@
+.tooltip {
+  width: 100%;
+  height: 100%;
+  position: absolute;
+  font-size: 20px;
+  text-decoration: none;
+  text-shadow: none;
+  border: none;
+  user-select: none;
+}
+
+.move {
+  position: absolute;
+  bottom: 100%;
+  right: 100%;
+  cursor: pointer;
+}
+
+.move::after {
+  content: "move";
+  position: absolute;
+  bottom: 110%;
+  right: -50%;
+  width: max-content;
+  color: black;
+  padding: 0.2rem;
+  border-radius: 2px;
+  background-color: white;
+  opacity: 0;
+  transition: all 200ms ease-in;
+}
+
+.move:hover::after {
+  opacity: 1;
+}
+
+.delete {
+  position: absolute;
+  top: 100%;
+  left: 100%;
+  cursor: pointer;
+}
+.delete::after {
+  content: "delete";
+  position: absolute;
+  top: 110%;
+  right: -50%;
+  width: max-content;
+  color: black;
+  padding: 0.2rem;
+  border-radius: 2px;
+  background-color: white;
+  opacity: 0;
+  transition: all 200ms ease-in;
+}
+
+.delete:hover::after {
+  opacity: 1;
+}
+
+.hidden {
+  display: none;
+}
